requirements.txt
pandas>=1.3.0
lyricsgenius>=3.0.0
ai4bharat-transliteration>=0.1.0
groq>=0.4.0
numpy>=1.20.0
matplotlib>=3.4.0
tqdm>=4.62.0
scikit-learn>=1.0.0
transformers>=4.12.0
torch>=1.10.0